BigCommerce Responsive Web Design Application
This is a book  Information Display Application that has been designed
and developed using Html5,Css3,Javascript,Ajax and RWS Approach.
Please Note :Resolution used are:320px,480px,1024px  and 1280px.
This application should be made to run on  local server as it is using
Ajax call methodology.
Also the browsers supported are IE9,10,Edge ,chrome and Firefox